requirements.txt
Django
djangorestframework
django-cors-headers
Pillow
django-colorfield
python-magic
xlrd
celery
django-celery-results
xlsxwriter